+++ functions/pagination.php
@@ -4,8 +4,8 @@
  *
  * @Author: Timi Wahalahti
  * @Date:   2020-01-10 15:47:21
- * @Last Modified by:   Timi Wahalahti
- * @Last Modified time: 2020-02-11 11:54:08
+ * @Last Modified by:   Elias Kautto
+ * @Last Modified time: 2021-12-17 11:08:17
  *
  * @package air-helper
  */
@@ -23,14 +23,6 @@ if ( ! function_exists( 'get_next_page_id' ) ) {
       $id = get_the_id();
     }
 
-    $cache_key = apply_filters( 'get_next_page_id_cache_key', "get_next_page_id_for_{$id}", $id );
-
-    // Check if result is cached and if, return cached version
-    $next_page_id = get_transient( $cache_key );
-    if ( ! empty( $next_page_id ) ) {
-      return absint( $next_page_id );
-    }
-
     $next_page_id = false;
 
     // Get all pages under this section
@@ -39,7 +31,7 @@ if ( ! function_exists( 'get_next_page_id' ) ) {
       'post_type'   => $post->post_type,
       'child_of'    => $post->post_parent,
       'parent'      => $post->post_parent,
-      'sort_column' => 'menu_order post_title',
+      'sort_column' => 'menu_order,post_title',
       'sort_order'  => 'ASC',
     ] );
 
@@ -61,9 +53,6 @@ if ( ! function_exists( 'get_next_page_id' ) ) {
       $next_page_id = $get_pages[ $next_key ]->ID;
     }
 
-    // Save to cache
-    set_transient( $cache_key, $next_page_id, apply_filters( 'get_next_page_id_cache_lifetime', MINUTE_IN_SECONDS * 30 ) );
-
     return $next_page_id;
   } // end get_next_page_id
 } // end if
@@ -81,14 +70,6 @@ if ( ! function_exists( 'get_prev_page_id' ) ) {
       $id = get_the_id();
     }
 
-    $cache_key = apply_filters( 'get_prev_page_id_cache_key', "get_prev_page_id_for_{$id}", $id );
-
-    // Check if result is cached and if, return cached version
-    $prev_page_id = get_transient( $cache_key );
-    if ( ! empty( $prev_page_id ) ) {
-      return absint( $prev_page_id );
-    }
-
     $prev_page_id = false;
 
     // Get all pages under this section
@@ -97,7 +78,7 @@ if ( ! function_exists( 'get_prev_page_id' ) ) {
       'post_type'   => $post->post_type,
       'child_of'    => $post->post_parent,
       'parent'      => $post->post_parent,
-      'sort_column' => 'menu_order post_title',
+      'sort_column' => 'menu_order,post_title',
       'sort_order'  => 'ASC',
     ] );
 
@@ -120,9 +101,39 @@ if ( ! function_exists( 'get_prev_page_id' ) ) {
       $prev_page_id = $get_pages[ $prev_key ]->ID;
     }
 
-    // Save to cache
-    set_transient( $cache_key, $prev_page_id, apply_filters( 'get_prev_page_id_cache_lifetime', MINUTE_IN_SECONDS * 30 ) );
-
     return $prev_page_id;
   } // end get_prev_page_id
+} // end if
+
+if ( ! function_exists( 'get_first_page_id' ) ) {
+  /**
+   *  Get ID of first page.
+   *
+   *  @since  2.13.0
+   *  @param  integer $id Page from which to count first to get, defaults to current page.
+   *  @return mixed       False if no first page, id if there is.
+   */
+  function get_first_page_id( $id = 0 ) {
+    if ( empty( $id ) ) {
+      $id = get_the_id();
+    }
+
+    $first_page_id = false;
+
+    // Get all pages under this section
+    $post = get_post( $id );
+    $get_pages = get_pages( [
+      'post_type'   => $post->post_type,
+      'child_of'    => $post->post_parent,
+      'parent'      => $post->post_parent,
+      'sort_column' => 'menu_order,post_title',
+      'sort_order'  => 'ASC',
+    ] );
+
+    if ( is_array( $get_pages ) ) {
+      $first_page_id = $get_pages[0]->ID;
+    }
+
+    return $first_page_id;
+  } // end get_first_page_id
 } // end if